#Neural_Networks : Neural Networks And Deep Learning
This Website is basically a book, regarding to Neural Networks and deep
learning (http://neuralnetworksanddeeplearning.com/index.html), written
by Michael Nielsen, and consists of a lot of basic material, explained
in excruciating detail and clarity. Must be read, because it will give
you a feel for the entire subject, as well as a roadmap for studying
the subject.

1) Read only the first three chapters as they contain all the working
knowledge required to understand the basics and the applications of
neural networks. The next two chapters are a bit technical, and should
be read after covering the contents of the first three chapters in
greater detail
2) Write down working programs for Chapter 1 and chapter 2, as these
two chapters form the crux of the subject, and should be understood
completely.
3)Understanding the Contents of Chapter 1 and 2 is essential, and
should be understood with immense clarity.
4)Go through the programs in chapter 3 loosely, since this chapter
points out various downfalls of the previous code, suggests further
optimisation algorithms and readings. The contents of this chapter is
important from a subjective point of view, since it will provide a
roadmap for future study, but the programs should be kept for tackling
later.