#Neural_Networks : Cauchy-Schwarz Inequality
While going through the chapter 1 of the book Neural Networks and deep
learning by Michael Nielsen (link provided in the previous commit),
there is a portion where Cauchy-Schwarz inequality is used to prove
that using gradient descent algorithm will always decrease the value of
the cost function. Here are the links which i referred to for
understanding the Cauchy-Schwarz inequality.

1)wikipedia(this link explains the basic form and gives a short proof
for the Cauchy-Schwarz inequality) :
https://en.wikipedia.org/wiki/Cauchy%E2%80%93Schwarz_inequality

2)Khan academy(this video gives a complete proof of the Cauchy-Schwarz
inequality, and explains the use of this theorem) :
https://www.khanacademy.org/math/linear-algebra/vectors_and_spaces/dot_c
ross_products/v/proof-of-the-cauchy-schwarz-inequality

After referring to these two links (that is if you aren’t already
familiar with it), you should try to solve the exercise given in the
book, to prove that the gradient descent algorithm always reduces the
value of the quadratic cost function by using the Cauchy-Schwarz
inequality