#Neural_Networks : Neural Networks and Machine Learning
I have started studying the subject of neural networks, after going
through the book mentioned in the earlier commit, by referring the book
- Neural networks and machine Learning (3rd Edition) by Simon Haykin.
(PDF LINK :
http://www.mif.vu.lt/~valdas/DNT/Literatura/Haykin09/Haykin09.pdf)
(Though I really advice you that you get a hard copy because you will
have to spend a lot of time (and I mean A LOT) with this book)

This book consists of a lot of subject matter, and I am eventually
planning to study this book along with the MIT’s machine learning
course 6.867, which covers somewhat similar subject matter (but for
now, for the basics, I’ll just stick with Simon Haykin).

This book is a little tedious, and requires a lot of understanding and
searching a lot of places for complementary material, since a lot of
things are considered to be prerequisites and thus only mentioned or
explained vaguely.
So, I will keep posting links, as a supplement material for the
portions of the book that I found difficult to understand personally.