#Neural_Networks : Demonstration of Perceptron Convergence Theorem
This youtube video demonstrates the working of basic Perceptron
Network, implementing the Perceptron convergence algorithm, to classify
a linearly separable dataset into two classes.
https://www.youtube.com/watch?v=vGwemZhPlsA

After reading the perceptron convergence theorem in the book Neural
Networks and Learning Machines - Simon Haykin, you should clearly
understand what is going on in this video, and if not, you should (as I
did) watch this video at a slower speed and try to understand what
exactly is going on.