#Neural_Networks : Bayes Theorem
In the book Neural Networks and Learning Machines (3rd Edition), Pg:55,
there is a comparison between a Perceptron classifier and the Bayes
Classifier. Not knowing what the Bayes’ classifier was, I tried to look
it up online and stumbled upon the Bayes’ theorem.The Bayes Classifier
is based on the famous Bayes’ Theorem, put forward by Thomas Bayes.
The Bayes’ Theorem intrigued some of my curiosity and hence I started
to search for resources that would explain it the best to me.

The links below give some beautiful explanation (both intuitive and
technical) of the Bayes’ theorem, and its implications.

intuitive explanation :
http://www.yudkowsky.net/rational/bayes

technical explanation :
http://www.yudkowsky.net/rational/technical

Even though the discussion given in the above links is mostly off topic
and has little to do with the Bayes’ classifier, I personally found it
very interesting and recommend reading it.

Also, the website (http://www.yudkowsky.net) contains a lot of other
blogs, which are also very interesting and one should give it a try.