#Neural_Networks : Interesting Artificial Intelligence Websites
In my earlier commit, I mentioned the site (http://www.yudkowsky.net).
This site is basically a blog by Eliezer S. Yudokowsky, a research
fellow at the Machine Intelligence Research Institute.

This redirected me to the Machine Intelligence Research Institute
website (https://intelligence.org).
The contents on this website are quiet interesting, especially the
Blog. Also, they have put their papers online, which lay out the
theoretical foundation of building aligned artificial intelligence
systems.

Also, through the website mentioned above, i found out another website
(http://lesswrong.com).
This website in a single word, is amazing.
The goal of this website being to explore the rationality of the human
brain, and relate it in terms of artificial intelligence. The blogs on
this website contain a lot of information and insights into a lot of
topics related to human brain and artificial intelligence. It is worth
bookmarking.