#Probability : MIT 6.041
As I described in the previous Neural_Networks commit, I am now
starting with learning probability theory through the MIT ocw course
6.041.
The book I’ll be referring to is the course textbook :
Introduction to Probability (2nd edition) - Bertsekas, Dimitri, and
John Tsitsiklis.
Even though the 3rd edition of the book is out, it does not contain any
practice problems. Hence, I’ll be referring to the 2nd edition.

My workflow for studying this course will be following :
1)Watch Lecture
2)Read course notes pertaining to the lecture
3)Read recitation notes and solve recitation problems
4)Solve assignment problems.